requirements.txt
PyMuPDF==1.24.1
Pillow==10.1.0
python-dotenv==1.0.0
google-generativeai==0.3.0
pydantic==2.5.0
